Split implementation of sliders to separate files. Get rid of CRTP as not longer needed.
