+++ source/gui/slider/Constants.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace gui::slider
+{
+    
+constexpr float THIN_TRACK_THICKNESS = 10.f;
+constexpr float THUMB_THICKNESS = 20.f;
+constexpr float BOUNDS_THICKNESS = 10.f; 
+
+} // namespace gui::slider
