+++ source/gui/slider/LayoutBuilder.cpp
@@ -1,0 +1,55 @@
+#include "gui/slider/LayoutBuilder.hpp"
+
+namespace gui::slider
+{
+
+gui::FramedSprite::LayoutConfig buildLayoutConfig(const sf::Vector2f& cornerSizes, const gui::FramedSprite::LayoutConfig::UVs& uvs)
+{
+    gui::FramedSprite::LayoutConfig layoutConfig{
+        .cornerSizes = 
+        {
+            .topLeft        = {cornerSizes.x, cornerSizes.y},
+            .bottomRight    = {cornerSizes.x, cornerSizes.y}
+        },
+        .uvs = uvs
+    };
+
+    return layoutConfig;
+} 
+
+gui::FramedSprite::LayoutConfig::UVs buildUVsForTrackTexture()
+{
+    return gui::FramedSprite::LayoutConfig::UVs
+    {
+        .topLeft        = {0.0f,   0.0f,  2.0f, 2.0f},
+        .topRight       = {4.0f,   0.0f,  2.0f, 2.0f},
+        .bottomLeft     = {0.0f,   4.0f,  2.0f, 2.0f},
+        .bottomRight    = {4.0f,   4.0f,  2.0f, 2.0f},
+    };
+}
+
+gui::FramedSprite::LayoutConfig buildLayoutConfigForTrackTexture()
+{
+    static auto layout = buildLayoutConfig({4.f, 4.f}, buildUVsForTrackTexture());
+    return layout;
+}
+
+
+gui::FramedSprite::LayoutConfig::UVs buildUVsForThumbTexture()
+{
+    return gui::FramedSprite::LayoutConfig::UVs
+    {
+        .topLeft        = {0.0f,   0.0f,  2.0f, 2.0f},
+        .topRight       = {4.0f,   0.0f,  2.0f, 2.0f},
+        .bottomLeft     = {0.0f,   4.0f,  2.0f, 2.0f},
+        .bottomRight    = {4.0f,   4.0f,  2.0f, 2.0f},
+    };
+}
+
+gui::FramedSprite::LayoutConfig buildLayoutConfigForThumbTexture()
+{
+    static auto layout = buildLayoutConfig({4.f, 4.f}, buildUVsForThumbTexture());
+    return layout;
+}
+
+}  // namespace gui::slider
