+++ source/gui/slider/VerticalThick.cpp
@@ -1,0 +1,11 @@
+#include "gui/slider/VerticalThick.hpp"
+
+namespace gui::slider
+{
+
+std::unique_ptr<VerticalThick> VerticalThick::create()
+{
+    return std::unique_ptr<VerticalThick>{new VerticalThick{}};
+}
+
+}  // namespace gui::slider
